feat(bet4-drain27): extract wiki tools into a scoped ToolPack (EP-8DC217EB BET-4)
Process-Spine-Decision: BET-4 drain-to-zero (plan §4) — behavior-identical pack extraction; ratchet freezes the shrunk inline set.
BI: BI-2B7EE073
